requirements.txt
streamlit
langchain
langchain-community
python-dotenv
requests
pypdf
python-docx
sentence-transformers
chromadb
faiss-cpu